requirements.txt
julep
requests
python-dotenv
pyyaml